Extract DefaultDict to main MCT folder
