+++ model_compression_toolkit/core/common/quantization/quantization_params_generation/qparams_weights_computation.py
@@ -17,7 +17,7 @@ from typing import Dict, Any, Tuple
 import numpy as np
 
 from model_compression_toolkit.logger import Logger
-from model_compression_toolkit.core.common.defaultdict import DefaultDict
+from model_compression_toolkit.defaultdict import DefaultDict
 from model_compression_toolkit.core.common.framework_info import FrameworkInfo
 from model_compression_toolkit.core.common.quantization.node_quantization_config import NodeWeightsQuantizationConfig
