+++ model_compression_toolkit/defaultdict.py
