+++ tests/pytorch_tests/model_tests/feature_models/gptq_test.py
@@ -16,7 +16,7 @@ import numpy as np
 import torch
 import torch.nn as nn
 
-from model_compression_toolkit.core import DefaultDict
+from model_compression_toolkit import DefaultDict
 from model_compression_toolkit.target_platform_capabilities.target_platform import QuantizationMethod
 from model_compression_toolkit.gptq.common.gptq_constants import QUANT_PARAM_LEARNING_STR, MAX_LSB_STR
 from tests.pytorch_tests.model_tests.base_pytorch_feature_test import BasePytorchFeatureNetworkTest
